# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Ilm School Portal API
  version: 1.0.0
  description: |-
    Multi-school (multi-tenant) school management API. Every role except
    `system_admin` is scoped to exactly one school (`schoolId` on the JWT);
    `system_admin` is a platform-level role with no school of its own and
    can act across every school.

    **Auth**: session is an httpOnly `token` cookie (JWT), set by
    `POST /auth/login` and cleared by `POST /auth/logout`. There is no
    bearer-token flow — the cookie is sent automatically by the browser on
    same-origin requests (`withCredentials: true` on the frontend's axios
    client). A request with no cookie, an invalid/expired token, or a
    revoked token (see below) gets `401`.

    **Session lifetime**: tokens are signed for 1 hour. Any authenticated
    request past the halfway point of that window silently reissues the
    cookie, so an active user is never hard-logged-out mid-session — only
    genuine inactivity lets it expire. That sliding renewal is itself capped
    at 12 hours from the original login (`sessionStartedAt` in the token),
    so a session can't be kept alive indefinitely just by staying active —
    after 12 hours the user has to log in again regardless of activity.
    Logging out invalidates the token server-side immediately (via a
    `token_version` check on every request), not just client-side — and
    because there's no per-device session tracking, logging out on one
...
    sessions.

...

...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /report-cards/terms:
    get:
      tags:
        - Report Cards
      summary: List school terms
      description: Admin and teacher only. Teachers use this to pick a term when creating a report card; admin also manages the term catalog.
      responses:
        '200':
          description: Terms
          content:
            application/json:
              schema:
                type: object
                properties:
                  terms:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        start_date:
                          type: string
                          format: date
                        end_date:
                          type: string
                          format: date
        '403':
          $ref: '#/components/responses/Forbidden'
      operationId: getReportCardsTerms
    post:
      tags:
        - Report Cards
      summary: Create a school term
      description: Admin only.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - start_date
                - end_date
              properties:
                name:
                  type: string
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Term created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  term:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      start_date:
                        type: string
                        format: date
                      end_date:
                        type: string
                        format: date
        '400':
          description: Missing name/dates, or end_date before start_date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: A term with this name already exists for the school
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: postReportCardsTerms
  '/report-cards/students/{studentId}':
    get:
      tags:
        - Report Cards
      summary: List a student's report cards
      description: |-
        Gated by `report_cards`. Visible to admin/teacher at the student's
        school (a teacher only if they teach the student) or that student's
        own parent — never the student themselves. Each report card includes
        its nested per-subject grades.
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Report cards
          content:
            application/json:
              schema:
                type: object
                properties:
                  reportCards:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        term_id:
                          type: integer
                        term_name:
                          type: string
                        term_start_date:
                          type: string
                          format: date
                        term_end_date:
                          type: string
                          format: date
                        created_at:
                          type: string
                          format: date-time
                        updated_at:
                          type: string
                          format: date-time
                        subjects:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: integer
                              subject_name:
                                type: string
                              grade:
                                type: string
                              comment:
                                type: string
                                nullable: true
        '403':
          description: 'Feature disabled, or caller doesn''t have access to this student'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: getReportCardsStudentsByStudentId
    post:
      tags:
        - Report Cards
      summary: Create a report card for a student
      description: |-
        Admin, or a teacher who teaches this student. One report card per
        student per term — use `PUT /report-cards/{id}` to edit an
        existing one instead.
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - term_id
                - subjects
              properties:
                term_id:
                  type: integer
                subjects:
                  type: array
                  items:
                    type: object
                    required:
                      - subject_name
                      - grade
                    properties:
                      subject_name:
                        type: string
                      grade:
                        type: string
                      comment:
                        type: string
      responses:
        '201':
          description: Report card created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  reportCard:
                    type: object
                    properties:
                      id:
                        type: integer
        '400':
          description: Missing term_id/subjects, an invalid subject, or an unknown term for this school
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Not admin/teacher, feature disabled, wrong school, or (teacher) not assigned to this student''s class'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A report card already exists for this student and term
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: postReportCardsStudentsByStudentId
  '/report-cards/{id}':
    put:
      tags:
        - Report Cards
      summary: Replace a report card's subjects
      description: Admin, or a teacher who teaches the student this report card belongs to. Replaces the full subject list rather than patching individual rows.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - subjects
              properties:
                subjects:
                  type: array
                  items:
                    type: object
                    required:
                      - subject_name
                      - grade
                    properties:
                      subject_name:
                        type: string
                      grade:
                        type: string
                      comment:
                        type: string
      responses:
        '200':
          $ref: '#/components/responses/MessageOk'
        '400':
          description: Missing/invalid subjects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Not admin/teacher, feature disabled, or (teacher) no longer assigned to this student''s class'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Report card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: putReportCardsById
    delete:
      tags:
        - Report Cards
      summary: Delete a report card
      description: Admin, or a teacher who teaches the student this report card belongs to.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/MessageOk'
        '403':
          description: 'Not admin/teacher, feature disabled, or (teacher) no longer assigned to this student''s class'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Report card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: deleteReportCardsById
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: httpOnly JWT set by POST /auth/login. See the top-level description for session lifetime/refresh/revocation behavior.
...
  responses:
    MessageOk:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
    Forbidden:
      description: 'Authenticated, but not permitted (wrong role, or outside the caller''s school)'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
...